fix: design inconsistencies
